message-service/requirements.txt
Flask==2.0.1
mysql-connector-python==8.0.27
Werkzeug==2.0.1
phonenumbers==8.13.29
